#10922 show error for sample
